requirements.txt
# external requirements
coverage==6.3.2
pytest-cov==3.0.0
scikit-learn==0.24.2
dataclasses>=0.6
pyyaml==6.0
marshmallow-dataclass==8.5.8
pandas>=1.1.5
hydra-core==1.1.1
dvc >= 2.0.0